cli_chatbot/requirements.txt
# Core dependencies
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0

# API clients
requests>=2.31.0
google-generativeai>=0.3.0

# Local AI models (optional - for offline mode)
transformers>=4.35.0
torch>=2.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
